Added contacts-controller
