+++ routes/api/contacts.js
@@ -1,86 +1,19 @@
 const express = require("express");
 
-const {
-  listContacts,
-  getContactById,
-  addContact,
-  updateContact,
-  removeContact,
-} = require("../../models/contacts");
-
 const router = express.Router();
 
-const { HttpError } = require("../../helpers");
+const {
+  getAllContacts,
+  getContactId,
+  postContact,
+  putContact,
+  deleteContact,
+} = require("../../controllers/contacts-controller");
 
-const Joi = require("joi");
-const contactShema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().email().required(),
-  phone: Joi.string().required(),
-});
-
-router.get("/", async (req, res, next) => {
-  try {
-    const result = await listContacts();
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.get("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await getContactById(contactId);
-    if (!result) {
-      throw HttpError(404, `No contact found with this ${contactId}`);
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = contactShema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const result = await addContact(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
-router.put("/:contactId", async (req, res, next) => {
-  try {
-    const { error } = contactShema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const { contactId } = req.params;
-    const result = await updateContact(contactId, req.body);
-    if (!result) {
-      throw HttpError(404, `No contact found with this ${contactId}`);
-    }
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
-router.delete("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await removeContact(contactId);
-    if (!result) {
-      throw HttpError(404, `No contact found with this ${contactId}`);
-    }
-    // res.status(204).send();
-    res.json({ message: "Сontact deleted" });
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/", getAllContacts);
+router.get("/:contactId", getContactId);
+router.post("/", postContact);
+router.put("/:contactId", putContact);
+router.delete("/:contactId", deleteContact);
 
 module.exports = router;
